backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
PyMySQL==1.1.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.0.1
cryptography==41.0.3
Werkzeug==2.3.7
SQLAlchemy==2.0.20
alembic==1.11.3
gunicorn==21.2.0
marshmallow==3.20.1
email-validator==2.0.0
